Move RecentScores to its own file
